+++ src/mask/mod.rs
@@ -1,4 +1,2 @@
 mod definition;
 pub use definition::*;
-
-mod swizzle;